Split dashboard service into focused modules
